Moved Logo.js out of icons subdirectory, close #3546
