+++ src/gui/base/Logo.js
@@ -1,6 +1,6 @@
 // @flow
 
-import {assertMainOrNodeBoot} from "../../../api/common/Env"
+import {assertMainOrNodeBoot} from "../../api/common/Env"
 
 assertMainOrNodeBoot()
