+++ src/gui/builtinThemes.js
@@ -4,7 +4,7 @@
  * @file color/theme definitions for default themes.
  */
 
-import {getLogoSvg} from "./base/icons/Logo"
+import {getLogoSvg} from "./base/Logo"
 import type {Theme, ThemeId} from "./theme"
 import {assertMainOrNodeBoot} from "../api/common/Env"
